Websocket library standard conformiti extension, refactoring and bug fixing. Tested fully against public echo servers. Modification and simplifications of the Actro libray according to the Websocket static library
Websocket Library:

Fixed a bugs in client handshake code:

1. the client needs to send an HTTP full URI and not a WS URI wehn we decide to put a full URI and not just the a path in the request.
2. The regex in the "Parse URI.vi" method has been improved for the dummy resource name "/"

Fixed a bug in the handshake client code whereby the check if we have a valid URI in the

Refactored  the client/server handshake code.

Now the server handshake checks if a resource is available as of the input array of resource and it answers with a 404 Not Found HTTP response according to the RFC 6455. The answer is handled by the client handshake code which throws an error.
Server handshake similarly handles themalformed client handshake by answering with a 400 Bad Request HTTP response according to RFC 6455. The response is handled by the client code throwing an error. We added a "ProcessHandshakeResponse.vi" that fully check the server response as of RFC 6455, before there was only a partial check obtained by reusing "the ProcessHandshakeRequest.vi"server code.